refactor: move SetupServerArgs into init/args.cpp
Move the SetupServerArgs implementation out of `init.cpp` into a dedicated translation unit and add it to the build.

This keeps `init.cpp` smaller and isolates server argument registration code in one place.
