+++ src/CMakeLists.txt
@@ -194,6 +194,7 @@ add_library(bitcoin_node STATIC EXCLUDE_FROM_ALL
   index/coinstatsindex.cpp
   index/txindex.cpp
   init.cpp
+  init/args.cpp
   kernel/chain.cpp
   kernel/checks.cpp
   kernel/coinstats.cpp
